refactor(dashboard): extract sidebar/tab ordering into useSidebarOrdering
Move the three user-defined ordering overlays (SessionBar tab order #4831,
sidebar repo + per-repo session orders #4832), their two server-switch
refresh effects (#4831 / #4940), and the three reorder handlers out of
App.tsx into a dedicated hook. The activeServerId selector that only fed
those effects moves with them. Deps arrays and persist/load calls are
byte-identical.

Part of #5560 (App.tsx Shell decomposition).
